+++ node-api/src/main/kotlin/net/corda/nodeapi/config/SSLConfiguration.kt
@@ -7,6 +7,7 @@ interface SSLConfiguration {
     val keyStorePassword: String
     val trustStorePassword: String
     val certificatesDirectory: Path
-    val keyStoreFile: Path get() = certificatesDirectory / "sslkeystore.jks"
+    val sslKeystore: Path get() = certificatesDirectory / "sslkeystore.jks"
+    val nodeKeystore: Path get() = certificatesDirectory / "nodekeystore.jks"
     val trustStoreFile: Path get() = certificatesDirectory / "truststore.jks"
 }